move game score logic from App component to separate hook
